refactor: Move the rollout bucketing into PercentageRollout
GetNormalizedNumber sat on FeatureFlagEvaluator, which no longer has
anything to do with percentage rollouts now the v3 path is gone: its only
caller is PercentageByContextCondition. Moving it next to that caller
leaves the evaluator doing just slug lookup and delegation.

The method name is kept verbatim because the sibling provider libraries
implement the same function under the same name, and the test theory
moves across with its inputs and expected values untouched — they are
duplicated across every library to catch bucketing divergence.

Resolves the "Move to own class in BMBB-702" comment.
